Delete wfx_host_events header in favor of the new abstraction structure
