+++ examples/platform/silabs/BaseApplication.cpp
@@ -64,10 +64,9 @@
 #include <platform/silabs/platformAbstraction/SilabsPlatform.h>
 
 #ifdef SL_WIFI
-#include "WifiInterfaceAbstraction.h"
-#include "wfx_host_events.h"
 #include <app/clusters/network-commissioning/network-commissioning.h>
 #include <platform/silabs/NetworkCommissioningWiFiDriver.h>
+#include <platform/silabs/wifi/WifiInterfaceAbstraction.h>
 #endif // SL_WIFI
 
 #ifdef DIC_ENABLE
