+++ src/platform/silabs/CHIPDevicePlatformEvent.h
@@ -26,7 +26,7 @@
 
 #include <platform/CHIPDeviceEvent.h>
 #if CHIP_DEVICE_CONFIG_ENABLE_WIFI_STATION
-#include "wfx_host_events.h"
+#include <platform/silabs/wifi/WifiInterfaceAbstraction.h>
 #endif
 
 namespace chip {
