+++ src/platform/silabs/wifi/wf200/Wf200WifiInterface.cpp
@@ -29,7 +29,6 @@
 #include "sl_wfx_host.h"
 #include "sl_wfx_task.h"
 #include "task.h"
-#include "wfx_host_events.h"
 #include <lib/support/CHIPMemString.h>
 #include <lib/support/CodeUtils.h>
 #include <lib/support/logging/CHIPLogging.h>
@@ -1219,7 +1218,7 @@ void wfx_dhcp_got_ipv4(uint32_t ip)
      */
     uint8_t ip4_addr[4];
 
-    ip4_addr[0] = (ip) &0xFF;
+    ip4_addr[0] = (ip) & 0xFF;
     ip4_addr[1] = (ip >> 8) & 0xFF;
     ip4_addr[2] = (ip >> 16) & 0xFF;
     ip4_addr[3] = (ip >> 24) & 0xFF;
